+++ doc/tutorial/pairwise_alignment/pairwise_alignment_solution_3.cpp
@@ -13,7 +13,7 @@ int main()
     auto seq2 = "AFLPGWQEENKLSKIWMKDCGCLW"_aa27;
 
     // Configure the alignment kernel.
-    auto config = seqan3::align_cfg::mode{seqan3::global_alignment} |
+    auto config = seqan3::align_cfg::method_global |
                   seqan3::align_cfg::scoring{seqan3::aminoacid_scoring_scheme{
                       seqan3::aminoacid_similarity_matrix::BLOSUM62}} |
                   seqan3::align_cfg::gap{seqan3::gap_scheme{seqan3::gap_score{-2}, seqan3::gap_open_score{-9}}};