+++ test/performance/alignment/global_affine_alignment_benchmark.cpp
@@ -26,7 +26,7 @@
     #include <seqan/align.h>
 #endif
 
-constexpr auto affine_cfg = seqan3::align_cfg::mode{seqan3::global_alignment} |
+constexpr auto affine_cfg = seqan3::align_cfg::method_global |
                             seqan3::align_cfg::gap{seqan3::gap_scheme{seqan3::gap_score{-1},
                                                                       seqan3::gap_open_score{-10}}} |
                             seqan3::align_cfg::scoring{seqan3::nucleotide_scoring_scheme{seqan3::match_score{4},